+++ tests/label-creation/usecases/delete-label.interactor.spec.ts
@@ -16,7 +16,6 @@ describe("DeleteLabelInteractor", () => {
 
    const deleteLabelPresenter: jest.Mocked<IDeleteLabelPresenter> = {
       present: jest.fn(),
-      setCallback: jest.fn(),
    };
 
    const deleteLabelValidator: jest.Mocked<IDeleteLabelValidation> = {