+++ tests/todo-modification/usecases/mark-todo-as-completed.interactor.spec.ts
@@ -34,7 +34,6 @@ describe("MarkTodoAsCompletedInteractor", () => {
 
    const presenter: jest.Mocked<IMarkTodoAsCompletedPresenter> = {
       present: jest.fn(),
-      setCallback: jest.fn(),
    };
 
    const validator: jest.Mocked<IMarkTodoAsCompletedValidation> = {