+++ tests/todo-modification/usecases/mark-todo-as-uncompleted.interactor.spec.ts
@@ -35,7 +35,6 @@ describe("MarkTodoAsUncompletedInteractor", () => {
 
    const presenter: jest.Mocked<IMarkTodoAsUncompletedPresenter> = {
       present: jest.fn(),
-      setCallback: jest.fn(),
    };
 
    const validator: jest.Mocked<IMarkTodoAsUncompletedValidation> = {