+++ CHANGELOG.md
@@ -11,8 +11,8 @@
 ### Minor Changes
 
 - 339af0c: - create label interface implementation for interactor and validation
-  - delete label interface implementation for interactor and validation
-  - get all label interface implementation for interactor
+   - delete label interface implementation for interactor and validation
+   - get all label interface implementation for interactor
 
 ## 1.7.1